+++ Web/Memorial/Core/Dtos/UrnDto.cs
@@ -17,7 +17,7 @@ namespace Memorial.Core.Dtos
 
         public SiteDto SiteDto { get; set; }
 
-        public byte SiteDtoId { get; set; }
+        public int SiteDtoId { get; set; }
 
         public DateTime CreateDate { get; set; }
     }
